UPDATE: Moved SendService to HttpsHandler. All new messages get echoed by the bot
