+++ TelegramIO/src/main/java/TelegramSmartHome/TelegramIO/UpdateService.java
@@ -1,5 +1,6 @@
 package TelegramSmartHome.TelegramIO;
 
+import TelegramSmartHome.TelegramIO.HttpsApiComm.HttpsHandler;
 import TelegramSmartHome.TelegramIO.NewMessageHandler.JsonHandler;
 import TelegramSmartHome.TelegramIO.NewMessageHandler.Update;
 import  io.vavr.control.Try;
@@ -12,17 +13,20 @@ public class UpdateService {
     private long lastUpdateId;
     List<IMessageEvaluator> evaluators;
     private JsonHandler jsonHandler;
+    private HttpsHandler httpsHandler;
 
     public UpdateService(String botToken){
         this.token=botToken;
-        jsonHandler = new JsonHandler(botToken);
+        this.httpsHandler = new HttpsHandler(this.token);
+        jsonHandler = new JsonHandler(botToken, httpsHandler);
         evaluators = new ArrayList<>();
     }
 
     public UpdateService(String botToken, int lastUpdateId){
         this.token=botToken;
+        this.httpsHandler = new HttpsHandler(this.token);
         this.lastUpdateId = lastUpdateId;
-        jsonHandler = new JsonHandler(botToken);
+        jsonHandler = new JsonHandler(botToken, httpsHandler);
         evaluators = new ArrayList<>();
 
     }
@@ -50,7 +54,14 @@ public class UpdateService {
     private void getUpdates(){
         List<Update> newMessages = jsonHandler.getNewMessages(lastUpdateId);
         newMessages.forEach(this::notifyUpdateListeners);
+        //Just echoes all new Messages 
+        newMessages.forEach(this::echoInput);
         lastUpdateId = (newMessages.size() > 0) ? newMessages.get(newMessages.size()-1).getUpdate_id() : lastUpdateId;
+    }
+
+    private void echoInput(Update update) {
+        long chatId = update.getMessage().getChat().getId();
+        httpsHandler.httpsPushRequest(chatId, update.getMessage().getMessageText());
     }
 
     private void notifyUpdateListeners(Update update) {
