+++ TelegramIO/src/test/java/TelegramSmartHome/TelegramIO/NewMessageHandlerTest/JsonHandlerTest.java
@@ -1,5 +1,6 @@
 package TelegramSmartHome.TelegramIO.NewMessageHandlerTest;
 
+import TelegramSmartHome.TelegramIO.HttpsApiComm.HttpsHandler;
 import TelegramSmartHome.TelegramIO.NewMessageHandler.*;
 import org.junit.BeforeClass;
 import org.junit.Test;
@@ -7,7 +8,6 @@ import org.junit.Test;
 import java.util.List;
 
 import static org.junit.Assert.assertEquals;
-import static org.mockito.ArgumentMatchers.anyLong;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
 
@@ -16,6 +16,7 @@ public class JsonHandlerTest {
     static String testJson;
     static String testSecondJson;
     static JsonHandler jsonHandler;
+    static HttpsHandler httpsHandler;
     static List<Update> newMessages;
     static Message expectedMessage;
 
@@ -23,9 +24,9 @@ public class JsonHandlerTest {
     public static void before() {
         testJson = "{\"ok\":true,\"result\":[{\"update_id\":1,\"message\":{\"message_id\":2,\"from\":{\"id\":42,\"is_bot\":false,\"first_name\":\"Test\",\"last_name\":\"Test\",\"username\":\"TestUser\",\"language_code\":\"en\"},\"chat\":{\"id\":42,\"first_name\":\"Test\",\"last_name\":\"Test\",\"username\":\"TestUser\",\"type\":\"private\"},\"date\":1563531123,\"text\":\"Hallp\"}}]}";
         testSecondJson = "{\"ok\":true,\"result\":[{\"update_id\":2,\"message\":{\"message_id\":3,\"from\":{\"id\":42,\"is_bot\":false,\"first_name\":\"Test\",\"last_name\":\"Test\",\"username\":\"TestUser2\",\"language_code\":\"en\"},\"chat\":{\"id\":42,\"first_name\":\"Test\",\"last_name\":\"Test\",\"username\":\"TestUser2\",\"type\":\"private\"},\"date\":1563531123,\"text\":\"User2\"}}]}";
-        jsonHandler = mock(JsonHandler.class);
-        when(jsonHandler.httpsGetRequest()).thenReturn(testJson).thenReturn(testSecondJson);
-        when(jsonHandler.getNewMessages(anyLong())).thenCallRealMethod();
+        httpsHandler = mock(HttpsHandler.class);
+        when(httpsHandler.httpsGetRequest()).thenReturn(testJson).thenReturn(testSecondJson);
+        jsonHandler = new JsonHandler("token", httpsHandler);
 
         newMessages = jsonHandler.getNewMessages(0);
